requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
sse-starlette==2.1.0
anthropic>=0.42.0
supabase>=2.28.0
boto3==1.35.0
pdfplumber==0.11.4
PyPDF2>=3.0.0
python-pptx==1.0.2
python-docx==1.1.2
openpyxl==3.1.5
httpx>=0.24.0,<1
pyjwt[crypto]>=2.10.1
svix==1.60.0
